Start cleaning up linter errors. (#358)
* Start cleaning up linter errors. For #357

* Fix unmarshalling NullTime values

* More linter fixes

* Remove commented code

* Move defer up

* Consolidate error check lines

* Move error check to make sure row iteration was successful

* Cleaner error check + do not recreate pipe if it exists

* Consolidate hashing to generate client id

+++ metrics/alerting.go
@@ -51,5 +51,5 @@ func handleDiskAlerting() {
 }
 
 func recentAverage(values []timestampedValue) int {
-	return int((values[len(values)-1].Value + values[len(values)-2].Value) / 2)
+	return (values[len(values)-1].Value + values[len(values)-2].Value) / 2
 }
